moved TypeDetail.h back to its original place
